Update package version and fix issue with displaying sale payment method type. (#123)
Bump the package version in `package.json` to `0.0.6`.

In `BuscarVenda.vue`, fix an issue where the sale payment method type was not displayed correctly. The issue was resolved by applying optional chaining to `sale.saleMethodPayment[0]?.methodPayment?.tipo?.nome`.

This change is necessary to ensure that the correct payment method type is shown for each sale, even when there are multiple payment methods.

Fixes #123.